+++ components/vf-card/vf-card.react.js
@@ -3,9 +3,10 @@
 // We use vanilla JS templates for react for compatibility with create react app
 // ---
 import React from "react";
-import vfNunjucks from "@visual-framework/vf-extensions-react/assets/nunjucks-slim.js";
-import VfCardTemplate from "raw-loader!./vf-card.precompiled.js"; // https://webpack.js.org/loaders/raw-loader/
-eval(VfCardTemplate); // we use eval as we specifically want to run a known template
+import Fragment from "react-dom-fragment";
+// eslint-disable-next-line no-unused-vars
+import VfCardTemplate from "./vf-card.precompiled.js"; // import templates before the nunjucks env
+import { vfNunjucksEnv } from "@visual-framework/vf-extensions-react/vf-extensions-react.js";
 
 // any JS actions needed on component insertion
 class VfCardCallback extends React.Component {
@@ -14,7 +15,7 @@ class VfCardCallback extends React.Component {
   }
 
   render() {
-    return React.createElement("div", null);
+    return React.createElement(React.Fragment, null);
   }
 }
 
@@ -22,11 +23,11 @@ const VfCard = React.memo(({
   variant, newTheme, card_image, card_text, card_image__alt, card_title
 }) => {
   // our HTML is handled by nunjucks, this should not receive user input
-  return React.createElement("div", null,
-    React.createElement("div", {
+  return React.createElement(React.Fragment, null,
+    React.createElement(Fragment, {
       dangerouslySetInnerHTML: {
         // our HTML is handled by nunjucks, this should not receive user input
-        __html: vfNunjucks.render("vf-card", {
+        __html: vfNunjucksEnv.render("vf-card", {
           variant: variant, newTheme: newTheme, card_image: card_image, card_text: card_text, card_image__alt: card_image__alt, card_title: card_title
         })
       }