+++ components/vf-button/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.3.1
+
+* React: Use react-dom-fragment to return HTML fragments.
+  * https://github.com/visual-framework/vf-core/pull/1291
+
 ### 1.3.0
 
 * adds react support