+++ tools/vf-extensions-react/CHANGELOG.md
@@ -1,3 +1,11 @@
+### 0.0.2
+
+* Use react-dom-fragment to return HTML fragments.
+  * https://github.com/visual-framework/vf-core/pull/1291
+* Removes no longer needed raw-loader
+* Add vf-extensions-react.js for common Nunjucks environment
+* VF Nunjucks extensions support
+
 ### 0.0.1
 
-* Initial release
+* Initial release.